* bug fix: rake would jump when it was around 360==0 degrees * also, changed method of smoothing rake to be more like smooth stroke ** may expose variables for smoothing rake later

+++ source/blender/makesdna/DNA_brush_types.h
@@ -93,7 +93,7 @@ typedef struct Brush {
 	float unprojected_radius;
 
 	/* record movement of mouse so that rake can start at an intuitive angle */
-	int last_x, last_y;
+	float last_x, last_y;
 	float last_angle;
 
 	int pad2;
